fix(images): Fix inabiltiy to find depreacted images by name
Currently it's not possible to query e.g. the `fedora-35` image by name,
after it was marked as deprecated due to the release of the `fedora-36`
image. Instead the library will treat this request as if the image
doesn't exist anymore.

This patch enables the `IncludeDeprecated` switch in the `GetByName`
function to allow images that were marked as deprecated to be found by
name.

This is necessary to allow e.g. the Terraform module, build on top of
this library, to continue to provide old OS versions for a transition
time without breaking.

References:
https://github.com/hetznercloud/terraform-provider-hcloud/blob/1dbb289398f19ca52b6b772501630381fc601fa0/internal/server/resource.go#L230
https://github.com/hetznercloud/terraform-provider-hcloud/issues/527