feat: Use generics to get pointers to types
Since the minimum required Go version is 1.18 according to go.mod we can
use generics to have a single function that can return a pointer to any
type. This avoids the need to add more $Type helpers in the future.

The new Ptr function is introduced and exsiting functions are updated to
call through Ptr instead. Any exported functions are now marked as
deprecated but not removed in order to avoid breaking backwards
compatibility. All non-exported functions are removed. All call sites
are updated to use Ptr directly.