feat(primary-ip): implement RDNSSupporter
Implement the RDNSSupporter interface so the reverse dns for primary
ips can be set the same way as for other resources (Server, Floating
IP, Load Balancer).

This is required for [terraform-provider-hcloud#668](https://github.com/hetznercloud/terraform-provider-hcloud/issues/668).

The existing struct `hcloud.PrimaryIPChangeDNSPtrOpts` has a bug, where
the DNSPtr field should be a `*string` to allow removing the entry, but
changing it is not possible without breaking backwards compatibility.
To work around this limitation, I added a new
`schema.PrimaryIPActionChangeDNSPtrRequest` that has the string pointer
field instead.