feat(servertype): implement new Deprecation api field (#268)
* feat(servertype): implement new Deprecation api field

We recently added a new Field to the ServerType API response that
indicates if and when the server type was deprecated. The same structure
will also be used for other resources in the future, replacing the
current "deprecated" fields.

* refactor(rdns): make sure the all expected resources comply with interface

* test: fix linting issue

Argument layout of mustParseTime was
always set to the same value.