feat: add interfaces for client structs (#342)
This commit adds auto-generated interfaces for client structs. This makes
mocking the clients easier in testing.

The interfaces have the same name as the corresponding struct, except for
beginning with an `I`. So for example, the interface for`ServerClient` would
be `IServerClient`.

The SDK itself will not use these new interface types, as most client structs
have exported properties and using the interface types would thus be a
breaking change.