fix: make schemas consistent with API (#348)
In some places the schema definitions are not consistent with the API
documentation, for example properties may not have pointer receivers
even though they are marked as optional in the documentation. This
results in inconsistent JSON when converting `JSON -> Schema -> JSON`
(as seen for example in https://github.com/hetznercloud/cli/issues/461,
where one output is the exact API response and one is serialized and
then converted back to JSON).

This PR aims to fix this issue by making the schema definitions
consistent.