fix: LoadBalancerUpdateServiceOpts not converted correctly (#394)
In the conversion from LoadBalancerUpdateServiceOpts to
schema.LoadBalancerActionUpdateServiceRequest, there are conversions
from slices to slice pointers. Slice pointers exist in schemas to
differentiate between absent and empty.

Goverter converts nil slices to &nil. This leads to the field being
marshaled to `null` in JSON (even with the `omitempty` flag) instead of
being not present (See https://github.com/hetznercloud/cli/issues/702)

This PR fixes this issue by adding manual conversion methods and adds
tests to account for this behavior.

---------

Co-authored-by: pauhull <22707808+pauhull@users.noreply.github.com>