fix: falsely omitted fields in firewall schema (#396)
In the [API Docs](https://docs.hetzner.cloud/#firewalls-get-a-firewall)
the `description`, `port`, `destination_ips` and `source_ips` properties
are not marked as required (thus being optional) although in practice
they are always set. This leads to inconsistencies between marshaled
schemas and API responses. This PR modifies the schema so that it is
consistent with the actual API output.