feat: implement actions waiter (#407)
Implement a simpler and more versatile waiting functions for actions.

Most use cases when waiting for actions is to return early if an action
fails. If all actions must be waited until completion, the users should
use the `WaitForFunc` function.

If the final actions objects are needed, the users should use the
`WaitForFunc` function to store the final actions using the
`handleUpdate` callback.

This deprecates the `ActionClient.WatchOverallProgress` and
`ActionClient.WatchProgress` methods.