feat: use custom IDOrName type for schemas (#545)
Add typing information to the fields that support passing either an ID
as int, an ID as string or a Name as string. An empty value will be
marshalled to `null` to reflect the behavior of an empty interface.

After this change, we do not have to cast IDs to float64 or to string to
be able to read the values.

Since this is part of the schema package, this is not a breaking change
for our customers, but might break our tests that rely on casting the
interface to the desired type.