fix: panic when a request did not set the OpPath for instrumentation
When an instrumented client (API or Metadata) made a request that did
not explicitly specify the OpPath through `ctxutil.SetOpPath()` that
instrumentation code caused a panic.

This is now handled by adding back a fallback if the explicit OpPath is
not set. The fallback uses the same patterns to create the OpPath as our
manual paths.

This only happened for the Metadata client and any custom
clients/requests that users added besides the official endpoints
implemented in hcloud-go.

Co-authored-by: Jonas Lammler <jonas.lammler@hetzner-cloud.de>